+++ src/chattool/tools/zulip/__init__.py
@@ -1,3 +1,0 @@
-from .client import ZulipClient
-
-__all__ = ["ZulipClient"]